requirements.txt
python-dotenv
PyPDF2
docx2txt
numpy
speechrecognition
gTTS
chromadb
google-generativeai
sentence-transformers
nltk
pyaudio
pygame 
pydub 
pyttsx3
webrtcvad
whisper
streamlit
elevenlabs